# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vahan-Rakshak API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://vahan-rakshak.onrender.com",
      "description": "Production server (Render)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/v1/driver/monitoring": {
      "post": {
        "summary": "Post Driver Monitoring",
        "operationId": "post_driver_monitoring_v1_driver_monitoring_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverMonitoringRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Post Driver Monitoring V1 Driver Monitoring Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/gatekeeper/run": {
      "post": {
        "summary": "Post Gatekeeper Run",
        "description": "Invoke the watsonx Gatekeeper agent with a specified action and payload.",
        "operationId": "post_gatekeeper_run_v1_gatekeeper_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatekeeperInvokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Post Gatekeeper Run V1 Gatekeeper Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/speed": {
      "post": {
        "summary": "Post Speed Reading",
        "operationId": "post_speed_reading_v1_speed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeedReadingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Post Speed Reading V1 Speed Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/cargo/scan-qr": {
      "post": {
        "summary": "Tool Cargo Scan Qr",
        "operationId": "tool_cargo_scan_qr_v1_tools_cargo_scan_qr_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargoScanQrRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Cargo Scan Qr V1 Tools Cargo Scan Qr Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/cargo/create-manifest": {
      "post": {
        "summary": "Tool Cargo Create Manifest",
        "operationId": "tool_cargo_create_manifest_v1_tools_cargo_create_manifest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargoCreateManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Cargo Create Manifest V1 Tools Cargo Create Manifest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/regulator/check-cargo-compliance": {
      "post": {
        "summary": "Tool Reg Check Cargo",
        "operationId": "tool_reg_check_cargo_v1_tools_regulator_check_cargo_compliance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatorCargoComplianceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Reg Check Cargo V1 Tools Regulator Check Cargo Compliance Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/regulator/check-weight": {
      "post": {
        "summary": "Tool Reg Check Weight",
        "operationId": "tool_reg_check_weight_v1_tools_regulator_check_weight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatorWeightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Reg Check Weight V1 Tools Regulator Check Weight Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/regulator/check-sensors": {
      "post": {
        "summary": "Tool Reg Check Sensors",
        "operationId": "tool_reg_check_sensors_v1_tools_regulator_check_sensors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatorSensorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Reg Check Sensors V1 Tools Regulator Check Sensors Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/regulator/verify-permit": {
      "post": {
        "summary": "Tool Reg Verify Permit",
        "operationId": "tool_reg_verify_permit_v1_tools_regulator_verify_permit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatorPermitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Reg Verify Permit V1 Tools Regulator Verify Permit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/regulator/report-violation": {
      "post": {
        "summary": "Tool Reg Report Violation",
        "operationId": "tool_reg_report_violation_v1_tools_regulator_report_violation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatorReportViolationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Tool Reg Report Violation V1 Tools Regulator Report Violation Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools/regulator/violations/{vehicle_id}": {
      "get": {
        "summary": "Tool Reg Get Violations",
        "operationId": "tool_reg_get_violations_v1_tools_regulator_violations__vehicle_id__get",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Tool Reg Get Violations V1 Tools Regulator Violations  Vehicle Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverMonitoringRequest": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id",
            "description": "Vehicle identifier"
          },
          "eye_closure_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Eye Closure Pct"
          },
          "blink_duration_ms": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Blink Duration Ms"
          },
          "yawning_rate_per_min": {
            "type": "number",
            "minimum": 0.0,
            "title": "Yawning Rate Per Min"
          },
          "steering_variability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Steering Variability"
          },
          "lane_departures": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Lane Departures"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "eye_closure_pct",
          "blink_duration_ms",
          "yawning_rate_per_min",
          "steering_variability",
          "lane_departures"
        ],
        "title": "DriverMonitoringRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}